now bork shuts down cleanly
What was previously causing a panic during shutdown was awaiting the async
frame of `Terminal.notifyDisplayEvents`. Somehow the await was causing a
wrong restart in the frame. Since we have already a few work arounds in
place to deal with async, we just stop awaiting the async frame and let
bork shutdown.